+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com/) and this project uses [semantic versioning](http://semver.org/).
 
 ### [Unreleased]
+### Added
+- Advanced Matchmaking with custom filters and user properties.
+
 ### Changed
 - Script runtime RPC and HTTP hook errors now return more detail when verbose logging is enabled.